# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ServerPe Back-end API
  version: "1.0.0"
  description: |
    Auto-generated OpenAPI spec for the ServerPe back-end (router summary).
servers:
  - url: http://localhost:8888
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    SecretKeyHeader:
      type: apiKey
      in: header
      name: x-secret-key
    ServerPeCookie:
      type: apiKey
      in: cookie
      name: serverpe_user_token
  schemas:
...
    PincodeDetailsRequest:
      type: object
      properties:
        pincode:
          type: string
      required:
        - pincode
    SendOtpRequest:
      type: object
      properties:
        mobile_number:
          type: string
      required:
        - mobile_number
    VerifyOtpRequest:
      type: object
      properties:
        mobile_number:
          type: string
        otp:
          type: string
      required:
        - mobile_number
        - otp
paths:
...
  /mockapis/serverpeuser/api/pincode-details:
    post:
      summary: Fetch pincode details (requires API & Secret keys)
      security:
        - ApiKeyHeader: []
        - SecretKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PincodeDetailsRequest'
      responses:
        '200':
          description: Pincode details
  /mockapis/serverpeuser/api/pincodes:
    get:
      summary: Get all pincodes (requires API & Secret keys)
      security:
        - ApiKeyHeader: []
        - SecretKeyHeader: []
      responses:
        '200':
          description: List of pincodes
  /mockapis/serverpeuser/send-otp:
    post:
      summary: Send OTP to mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP send status
  /mockapis/serverpeuser/verify-otp:
    post:
      summary: Verify OTP and set authentication cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: Verify result (may set cookie)
  /mockapis/serverpeuser/api-usage:
    get:
      summary: Get API usage for logged-in serverpe user (cookie JWT)
      security:
        - ServerPeCookie: []
      responses:
        '200':
          description: Usage info